database changes without decryption

+++ AirlineReservationSys/src/ARSModel/Clerk.java
@@ -6,7 +6,7 @@ public class Clerk extends User {
 		super();
 	}
 	
-	public Clerk(long personID,String password, String name, String surname, String email, String socialSecurityNo,
+	public Clerk(int personID,String password, String name, String surname, String email, String socialSecurityNo,
 			String phoneNo){
 		
 		super(personID,password, name, surname, email,socialSecurityNo,phoneNo);
